+++ src/algorithms/local_search/insertion_search.h
@@ -137,7 +137,12 @@ RouteInsertion compute_best_insertion_pd(const Input& input,
     }
 
     // Build replacement sequence for current insertion.
-    std::vector<Index> modified_with_pd({j});
+    std::vector<Index> modified_with_pd;
+    if (pickup_r <= end_d_rank) {
+      modified_with_pd.reserve(end_d_rank - pickup_r + 2);
+    }
+    modified_with_pd.push_back(j);
+
     Amount modified_delivery = input.zero_amount();
 
     // No need to use begin_d_rank here thanks to