+++ src/problems/tsp/tsp.cpp
@@ -36,10 +36,10 @@ UserCost compute_cost(const std::list<Index>& tour,
   return cost;
 }
 
-TSP::TSP(const Input& input, std::vector<Index> job_ranks, Index vehicle_rank)
+TSP::TSP(const Input& input, std::vector<Index>&& job_ranks, Index vehicle_rank)
   : VRP(input),
     _vehicle_rank(vehicle_rank),
-    _job_ranks(std::move(job_ranks)),
+    _job_ranks(job_ranks),
     _has_start(_input.vehicles[_vehicle_rank].has_start()),
     _has_end(_input.vehicles[_vehicle_rank].has_end()) {
 
@@ -47,6 +47,8 @@ TSP::TSP(const Input& input, std::vector<Index> job_ranks, Index vehicle_rank)
 
   // Pick ranks to select from input matrix.
   std::vector<Index> matrix_ranks;
+  matrix_ranks.reserve(_job_ranks.size() + 2);
+
   std::transform(_job_ranks.cbegin(),
                  _job_ranks.cend(),
                  std::back_inserter(matrix_ranks),