+++ src/structures/generic/undirected_graph.cpp
@@ -42,8 +42,8 @@ UndirectedGraph<T>::UndirectedGraph(const Matrix<T>& m) : _size(m.size()) {
 }
 
 template <class T>
-UndirectedGraph<T>::UndirectedGraph(std::vector<Edge<T>> edges)
-  : _edges{std::move(edges)} {
+UndirectedGraph<T>::UndirectedGraph(std::vector<Edge<T>>&& edges)
+  : _edges(edges) {
   for (auto const& edge : _edges) {
     Index first = edge.get_first_vertex();
     Index second = edge.get_second_vertex();