+++ src/structures/vroom/input/input.cpp
@@ -958,8 +958,8 @@ void Input::set_matrices(unsigned nb_thread) {
   };
 
   std::vector<std::thread> matrix_threads;
-
   matrix_threads.reserve(thread_profiles.size());
+
   for (const auto& profiles : thread_profiles) {
     matrix_threads.emplace_back(run_on_profiles, profiles);
   }